Created new controls for product details panel, added scanner hint
New control for Polishness check list has been created. It shows red check mark if something applies to the found company or gray X symbol otherwise.
Progress bar to show PL score and PL capital has been implemented as a new control. It additionally shows the % value as a text.
Scanner page now has a hint text that informs user how to use the app. It disappears when first barcode is scanned.
Disabled text scale factor for all views, so the layout of the app will not change when user changes the font size for his phone.